Adding btag config into the same NuNuProducer module, also changing intersection selection
